// src/interfaces/CadastroProfissionalInterface.ts
export interface CadastroProfissionalInterface {
    id: number; 
    nome: string; 
    celular: string; 
    email: string; 
    cpf: string; 
    dataNascimento: string; 
    cidade: string; 
    estado: string; 
    pais: string; 
    rua: string; 
    numero: string; 
    bairro: string; 
    cep: string; 
    complemento: string; 
    senha: string;
    salario:string;
}